requirements.txt
chromadb==1.2.1
Flask==3.1.2
Flask_Discord==0.1.69
Flask_MySQLdb==2.0.0
python-dotenv==1.1.1
Requests==2.32.5
